Player Stats Revamp: Revamped the way querying is performed
Split apart the processing into the Fragment and moved calling the asynctask into the activity. Also, a handler will be in place to send any information from the asynctask back to the activity. If it fails, the error message will be as usual handled and action bar updated. If it succeeds, it will convert into a PlayerReply object and pass it over to the fragment to process.

The fragment will then receive this and process it the same way as it did in the AsyncTask for now. It will be revamped too :D