Player Stats Revamp: Implemented RecyclerView usage in new PlayerInfo
Now it uses recyclerview. So far it looks like all abilities from the former listview implementation is implemented into the new recyclerview implementation. There are still a couple of bugs to iron out such as some text not being coloured as well as the tab layout not hiding on scroll.

Also, session data are taken out due to recyclerviews not allowing additional data to be implemented. This will probably be used somewhere else